# packages/api-client/openapi.json
{
...
  "info": {
    "contact": {},
    "description": "REST API for the Hà Tĩnh immersive tourism platform foundation.",
    "title": "Hà Tĩnh Immersive API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/hotspots": {
      "post": {
        "operationId": "createHotspot",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "id": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "payload": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "pitch": {
                    "example": 0,
                    "type": "number"
                  },
                  "sceneId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "status": {
                    "enum": ["draft", "published", "archived"],
                    "type": "string"
                  },
                  "type": {
                    "enum": ["information", "media", "audio", "external"],
                    "type": "string"
                  },
                  "yaw": {
                    "example": 180,
                    "type": "number"
                  }
                },
                "required": ["sceneId", "type", "yaw", "pitch", "payload"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "payload": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "pitch": {
                      "maximum": 90,
                      "minimum": -90,
                      "type": "number"
                    },
                    "sceneId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "type": {
                      "enum": ["information", "media", "audio", "external"],
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "yaw": {
                      "exclusiveMaximum": true,
                      "maximum": 360,
                      "minimum": 0,
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "sceneId",
                    "type",
                    "yaw",
                    "pitch",
                    "payload",
                    "status",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created immersive hotspot."
          }
        },
        "summary": "",
        "tags": ["admin-immersive"]
      }
    },
    "/api/v1/admin/hotspots/{id}": {
      "patch": {
        "operationId": "updateHotspot",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "payload": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "pitch": {
                    "example": 0,
                    "type": "number"
                  },
                  "status": {
                    "enum": ["draft", "published", "archived"],
                    "type": "string"
                  },
                  "type": {
                    "enum": ["information", "media", "audio", "external"],
                    "type": "string"
                  },
                  "yaw": {
                    "example": 180,
                    "type": "number"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "payload": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "pitch": {
                      "maximum": 90,
                      "minimum": -90,
                      "type": "number"
                    },
                    "sceneId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "type": {
                      "enum": ["information", "media", "audio", "external"],
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "yaw": {
                      "exclusiveMaximum": true,
                      "maximum": 360,
                      "minimum": 0,
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "sceneId",
                    "type",
                    "yaw",
                    "pitch",
                    "payload",
                    "status",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Updated immersive hotspot."
          }
        },
        "summary": "",
        "tags": ["admin-immersive"]
      }
    },
    "/api/v1/admin/scene-links": {
      "post": {
        "operationId": "createSceneLink",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "bidirectional": {
                    "example": true,
                    "type": "boolean"
                  },
                  "fromSceneId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "id": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "pitch": {
                    "example": 0,
                    "type": "number"
                  },
                  "sortOrder": {
                    "example": 0,
                    "type": "integer"
                  },
                  "toSceneId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "yaw": {
                    "example": 90,
                    "type": "number"
                  }
                },
                "required": ["fromSceneId", "toSceneId", "yaw", "pitch"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "bidirectional": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "fromDestinationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "fromSceneId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "pitch": {
                      "maximum": 90,
                      "minimum": -90,
                      "type": "number"
                    },
                    "sortOrder": {
                      "type": "integer"
                    },
                    "toDestinationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "toSceneId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "yaw": {
                      "exclusiveMaximum": true,
                      "maximum": 360,
                      "minimum": 0,
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "fromSceneId",
                    "toSceneId",
                    "fromDestinationId",
                    "toDestinationId",
                    "yaw",
                    "pitch",
                    "bidirectional",
                    "sortOrder",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created scene graph link."
          }
        },
        "summary": "",
        "tags": ["admin-immersive"]
      }
    },
    "/api/v1/admin/scene-links/{id}": {
      "delete": {
        "operationId": "deleteSceneLink",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted scene graph link."
          }
        },
        "summary": "",
        "tags": ["admin-immersive"]
      }
    },
...
    "/api/v1/destinations/{slug}": {
      "get": {
        "operationId": "getDestination",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "categoryId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "categoryLabel": {
                      "nullable": true,
                      "type": "string"
                    },
                    "coverImageUrl": {
                      "nullable": true,
                      "type": "string"
                    },
                    "coverMediaId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "defaultSceneId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "geoPoint": {
                      "nullable": true,
                      "properties": {
                        "latitude": {
                          "example": 18.3421,
                          "type": "number"
                        },
                        "longitude": {
                          "example": 105.9032,
                          "type": "number"
                        }
                      },
                      "required": ["latitude", "longitude"],
                      "type": "object"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "summary": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "name",
                    "summary",
                    "coverImageUrl",
                    "categoryLabel",
                    "status",
                    "description",
                    "categoryId",
                    "defaultSceneId",
                    "geoPoint",
                    "coverMediaId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Published destination detail."
          }
        },
        "summary": "",
        "tags": ["destinations"]
      }
    },
...
  },
  "servers": [],
...
}